添加了测试结果/测试集，并继续进行了源码的修改（注：当前测试基于Commit 9312a77进行）

+++ clang/lib/AST/ASTContext.cpp
@@ -259,7 +259,7 @@ RawComment *ASTContext::getRawCommentForDeclNoCacheImpl(
     return nullptr;
 
   // Decompose the end of the comment.
-  const unsigned CommentEndOffset =
+  const uint64_t CommentEndOffset =
       Comments.getCommentEndOffset(CommentBeforeDecl);
 
   // Get the corresponding buffer.
